requirements.txt
fastapi
uvicorn
gunicorn
insightface
onnx
onnxruntime
mxnet
ultralytics
opencv-python
sqlalchemy
psycopg2-binary
alembic
python-multipart
python-dotenv
passlib[argon2]
python-jose
numpy
scipy
scikit-learn
tqdm
loguru
